wrapping functions
Added wrapper functions for: transfer and transferFrom. 
Those functions were added because we've seen that for security purposes when a functions is called it triggers another call to another function that does the work. Doing this we're following the pattern designed by the guys of OpenZeppelin.
Also added increaseAllowance and decreaseAllowance